// packages/ui/src/components/Editor/constants/enums.ts
enum E_MARK {
	BOLD = 'BOLD',
	ITALIC = 'ITALIC',
	UNDERLINE = 'UNDERLINE',
	CODE = 'CODE',
	STRIKE_THROUGH = 'STRIKE_THROUGH',
}

enum E_TOOLBAR_EXCLUDE_BLOCK {
	IMAGE = 'IMAGE',
	LINK = 'LINK',
	TAG = 'TAG',
}

enum E_ELEMENT {
	PARAGRAPH = 'PARAGRAPH',
	HEADING_ONE = 'HEADING_ONE',
	HEADING_TWO = 'HEADING_TWO',
	HEADING_THREE = 'HEADING_THREE',
	BLOCK_QUOTE = 'BLOCK_QUOTE',
	NUMBERED_LIST = 'NUMBERED_LIST',
	BULLETED_LIST = 'BULLETED_LIST',
	LIST_ITEM = 'LIST_ITEM',
	IMAGE = 'IMAGE',
	LINK = 'LINK',
	TAG = 'TAG',
}

enum E_FORMAT_TYPE {
	MARK = 'MARK',
	BLOCK = 'BLOCK',
}

export { E_ELEMENT, E_FORMAT_TYPE, E_MARK, E_TOOLBAR_EXCLUDE_BLOCK };